+++ src/main/java/ch/njol/skript/expressions/ExprEnchantmentBonus.java
@@ -37,7 +37,7 @@ import ch.njol.skript.lang.util.SimpleExpression;
 import ch.njol.skript.log.ErrorQuality;
 import ch.njol.util.Kleenean;
 
-@Name("Enchant Prepare Enchantment Bonus")
+@Name("Enchantment Bonus")
 @Description("The enchantment bonus in an enchant prepare event. This represents the number of bookshelves affecting/surrounding the enchantment table.")
 @Examples({"on enchant:", 
 	"\tsend \"There are %enchantment bonus% bookshelves surrounding this enchantment table!\" to player"})
